Next exercise: The Aldous-Broder Algorithm
It starts by picking a random cell in the maze. Then, it jumps to a random neighbor of that cell.
If that cell has not yet been visited, it connects it to the previous cell and marks it as
visited. Then, it jumps to another random neighbor and does the same thing, continuously
jumping randomly until all cells have been visited, at which point the algorithm terminates.
This could take a very long time..